Use _lazy_format in all logging
