+++ numba/core/byteflow.py
@@ -2,12 +2,11 @@
 Implement python 3.8+ bytecode analysis
 """
 import dis
-from pprint import pformat
 import logging
 from collections import namedtuple, defaultdict, deque
 from functools import total_ordering
 
-from numba.core.utils import UniqueDict, PYVERSION, ALL_BINOPS_TO_OPERATORS
+from numba.core.utils import UniqueDict, PYVERSION, ALL_BINOPS_TO_OPERATORS, _lazy_pformat
 from numba.core.controlflow import NEW_BLOCKERS, CFGraph
 from numba.core.ir import Loc
 from numba.core.errors import UnsupportedError
@@ -70,15 +69,6 @@ class BlockKind(object):
 
     def __repr__(self):
         return "BlockKind({})".format(self._value)
-
-
-class _lazy_pformat(object):
-    def __init__(self, *args, **kwargs):
-        self.args = args
-        self.kwargs = kwargs
-
-    def __str__(self):
-        return pformat(*self.args, **self.kwargs)
 
 
 class Flow(object):
