+++ numba/core/compiler_machinery.py
@@ -300,7 +300,7 @@ class PassManager(object):
             name=f"{pss.name()} [{qualname}]",
             qualname=qualname,
             module=internal_state.func_id.modname,
-            flags=pformat(internal_state.flags.values()),
+            flags=utils._lazy_pformat(internal_state.flags.values()),
             args=str(internal_state.args),
             return_type=str(internal_state.return_type),
         )
